bug fixes and reorganize
