+++ demos/mesh_fitting/model.py
@@ -1,11 +1,11 @@
 import jax
 import jax.numpy as jnp
 import genjax
-import b3d.likelihoods as likelihoods
-import b3d.differentiable_renderer as rendering
+import b3d.chisight.dense.likelihoods as likelihoods
+import b3d.chisight.dense.differentiable_renderer as rendering
 import b3d
 from b3d import Pose
-from b3d.model import uniform_pose
+from b3d.modeling_utils import uniform_pose
 import rerun as rr
 import demos.differentiable_renderer.patch_tracking.demo_utils as utils
 
@@ -104,8 +104,8 @@ def get_rgb_only_model(renderer, initial_mesh):
     outlier_prob = 0.05
     def normalize(x):
         return x / jnp.sum(x)
-    likelihood = b3d.likelihoods.ArgMap(
-        b3d.likelihoods.get_uniform_multilaplace_rgbonly_image_dist_with_fixed_params(
+    likelihood = b3d.chisight.dense.likelihoods.ArgMap(
+        b3d.chisight.dense.likelihoods.get_uniform_multilaplace_rgbonly_image_dist_with_fixed_params(
             renderer.height, renderer.width, color_scale
         ),
         lambda weights, attributes:(
@@ -113,7 +113,7 @@ def get_rgb_only_model(renderer, initial_mesh):
             attributes
         )
     )
-    hyperparams = b3d.differentiable_renderer.DifferentiableRendererHyperparams(3, 1e-5, 1e-2, -1)
+    hyperparams = b3d.chisight.dense.differentiable_renderer.DifferentiableRendererHyperparams(3, 1e-5, 1e-2, -1)
     model = model_factory(
         renderer, likelihood, hyperparams, mindepth, maxdepth, 1,
         initial_mesh[0].shape[0], initial_mesh[1].shape[0]
