+++ demos/mesh_fitting/tessellation.py
