+++ demos/tessellation.py
@@ -2,7 +2,7 @@ import jax
 import jax.numpy as jnp
 import genjax
 import b3d
-import b3d.tessellation as t
+import demos.mesh_fitting.tessellation as t
 import b3d.utils as u
 import os
 import rerun as rr
